requirements.txt
streamlit
google-genai
pandas
python-docx
requests
streamlit-lottie